LineBoardYamanoteのiPad/SP分離
