+++ src/components/LineBoardYamanotePad/PadArch.tsx
